The exclusion rule is the reader's, and one candidate key space serves both sweeps
The rule was filed as the session's. Issue #491 states it in its own words, so it is the reader's, and a session attribution is the difference between a ruling that binds and one the next session may drop without asking.

`sheet.rs` and `tests/config.rs` were each building a list of every key event a sweep walks, and the second had already drifted from the first: it omitted `Backspace`, `Tab`, `Delete` and `Insert` alongside the function keys, while its own comment claimed the function keys were the whole difference. A gate whose candidate space is narrower than the keymap passes over the binding it cannot reach, and nothing would have said so. The list moves to `tests/support/mod.rs`, which both binaries already reach, and the comment describing the delta goes with the delta.

The sweep dedups through a set of discriminants rather than a linear scan restated as a closure, and reads its event through `let`-else, which is what the rest of this tree does and what clippy asks for over a nested `if let`.
