+++ SPEC.md
@@ -1141,7 +1141,7 @@ The count is of *configuration*, which is what the amendments turn on. `VIGIA_CO
 
 **Amended 2026-08-25 by [#306](https://github.com/breferrari/vigia/issues/306): the view toggles get their defaults in a second file, `~/.config/vigia/config`, and the count of *files* goes to two while the count of variables stays at one.** Asked for from the pane, alongside B16's pin: `r` and `s` decide what the body is made of, both start off, and a reader who always wants one file at a time or always wants the rail was pressing the key every time the pane opened.
 
-**Seven keys, and `follow` is excluded on purpose.** Four of them are the gestures sheet's `view` section once follow is taken out of it, `rail`, `single`, `staged` and `wrap`, which is a derivation rather than a list and is what made `staged` land with B17 rather than after it; the fifth is B21's `notes`, which `c` reaches from the sheet's `notes` section; and B18's `icons` and `links` reach no key at all. **A toggle absent from this file is absent by ruling or by oversight, and only a list tells them apart**, so every gesture that flips what the body is made of has a key here or an entry on the exclusion list carrying its reason, and `follow` is the list's one entry. *Ruled 2026-09-09, session.* **It answers nothing [#50](https://github.com/breferrari/vigia/issues/50) asks** — that row decides which way the toggle *ships*, the shipped default is unchanged and off, and this only lets a reader who has already decided stop saying so every session. And they have in common that they are pure preference: every combination is a legitimate pane. **I5 is why follow is not among them.** That invariant is *correct with zero interaction, auto-follows the newest change*, and a key that could turn it off would be configuring away an invariant rather than expressing a preference. A reader who wants it off presses `f`, which is a session's choice about a session.
+**Seven keys, and `follow` is excluded on purpose.** Four of them are the gestures sheet's `view` section once follow is taken out of it, `rail`, `single`, `staged` and `wrap`, which is a derivation rather than a list and is what made `staged` land with B17 rather than after it; the fifth is B21's `notes`, which `c` reaches from the sheet's `notes` section; and B18's `icons` and `links` reach no key at all. **A toggle absent from this file is absent by ruling or by oversight, and only a list tells them apart**, so every gesture that flips what the body is made of has a key here or an entry on the exclusion list carrying its reason, and `follow` is the list's one entry. *Ruled 2026-09-09, reader.* **It answers nothing [#50](https://github.com/breferrari/vigia/issues/50) asks** — that row decides which way the toggle *ships*, the shipped default is unchanged and off, and this only lets a reader who has already decided stop saying so every session. And they have in common that they are pure preference: every combination is a legitimate pane. **I5 is why follow is not among them.** That invariant is *correct with zero interaction, auto-follows the newest change*, and a key that could turn it off would be configuring away an invariant rather than expressing a preference. A reader who wants it off presses `f`, which is a session's choice about a session.
 
 **The cost is the honest measure of this amendment, and it is a file's worth of parsing and nothing else.** The discovery rule, the grammar, the comment syntax, the unknown-key refusal, the line-numbered errors and the report-before-the-takeover order all exist for the theme file, so what is added is a second place to look and keys to look for, exactly as the `VIGIA_GLYPHS` amendment above added a detector rather than a surface. Absent is not an error and unreadable is, which is the theme file's distinction for the theme file's reason.
